src/python/requirements.txt
google-generativeai
supabase
python-dotenv
colorama
requests
websockets>=13.0
